+++ ep-sdk/src/EpSdkLoggerCodes.ts
@@ -1,0 +1,4 @@
+export enum EEpSdkLoggerCodes {
+  SERVICE_GET = "SERVICE_GET",
+  SERVICE_CREATE = "SERVICE_CREATE",
+}