+++ ep-sdk/test/lib/TestConfig.ts
@@ -1,3 +1,4 @@
+import { EEpSdkLogLevel, EpSdkLogger } from "../../src/EpSdkLogger";
 
 enum EEnvVars {
   EP_SDK_TEST_SOLACE_CLOUD_TOKEN = 'EP_SDK_TEST_SOLACE_CLOUD_TOKEN',
@@ -36,7 +37,8 @@ class TestConfig {
 
     this.testConfig = {
       epBaseUrl: this.getOptionalEnvVarValueAsUrlWithDefault(EEnvVars.EP_SDK_TEST_EP_API_BASE_URL, this.DEFAULT_EP_SDK_TEST_EP_API_BASE_URL),
-    }
+    };
+
   }
 
   public getAppId = (): string => {