+++ ep-sdk/test/setup.spec.ts
@@ -4,7 +4,9 @@ import { expect } from 'chai';
 import { TestLogger } from "./lib/TestLogger";
 import TestConfig from "./lib/TestConfig";
 import { TestContext } from "./lib/TestContext";
-import { EPClient } from "../src/EPClient";
+import { EpSdkClient } from "../src/EpSdkClient";
+import { EpSdkConsoleLogger } from "../src/EpSdkConsoleLogger";
+import { EEpSdkLogLevel, EpSdkLogger } from "../src/EpSdkLogger";
 
 // load test stub
 const x = require('./lib/TestStub');
@@ -38,9 +40,11 @@ describe(`${scriptName}`, () => {
       TestContext.newItId();
     });
 
-    it(`${scriptName}: should initialize test config`, async () => {
+    it(`${scriptName}: should initialize test config & logger`, async () => {
       try {
         TestConfig.initialize();
+        const epSdkConsoleLogger: EpSdkConsoleLogger = new EpSdkConsoleLogger(TestConfig.getAppId(), EEpSdkLogLevel.Trace);
+        EpSdkLogger.initialize({ epSdkLoggerInstance: epSdkConsoleLogger });
       } catch (e) {
         expect(false, TestLogger.createTestFailMessageForError('intitializing test config failed', e)).to.be.true;
       }
@@ -48,7 +52,7 @@ describe(`${scriptName}`, () => {
 
     it(`${scriptName}: should initialize EP client`, async () => {
       try {
-        EPClient.initialize({
+        EpSdkClient.initialize({
           token: TestConfig.getSolaceCloudToken(),
           baseUrl: TestConfig.getConfig().epBaseUrl,
         });