+++ ep-sdk/test/specs/services/applicationDomainService.spec.ts
@@ -1,0 +1,113 @@
+import 'mocha';
+import { expect } from 'chai';
+import path from 'path';
+import { TestLogger } from '../../lib/TestLogger';
+import { TestContext } from '../../lib/TestContext';
+import TestConfig from '../../lib/TestConfig';
+import { TestUtils } from '../../lib/TestUtils';
+import { ApiError, ApplicationDomain, ApplicationDomainResponse, ApplicationDomainsService } from '../../../src/sep-openapi-node';
+import EpSdkApplicationDomainsService from '../../../src/services/EpSdkApplicationDomainsService';
+import { EpSdkError, EpSdkServiceError } from '../../../src/EpSdkErrors';
+
+const scriptName: string = path.basename(__filename);
+TestLogger.logMessage(scriptName, ">>> starting ...");
+
+
+const ApplicationDomainName = `${TestConfig.getAppId()}/${TestUtils.getUUID()}`;
+let ApplicationDomainId: string | undefined;
+
+describe(`${scriptName}`, () => {
+    
+    beforeEach(() => {
+      TestContext.newItId();
+    });
+
+    it(`${scriptName}: should create application domain`, async () => {
+      try {
+        const applicationDomainResponse: ApplicationDomainResponse = await ApplicationDomainsService.createApplicationDomain({
+          requestBody: {
+            name: ApplicationDomainName,
+          }
+        });
+        ApplicationDomainId = applicationDomainResponse.data.id;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should get application domain by name`, async () => {
+      try {
+        const applicationDomain: ApplicationDomain | undefined = await EpSdkApplicationDomainsService.getByName({ applicationDomainName: ApplicationDomainName });
+        // expect(false, TestLogger.createLogMessage('check logging')).to.be.true;
+        expect(applicationDomain, TestLogger.createApiTestFailMessage('applicationDomain === undefined')).to.not.be.undefined;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should get application domain by id`, async () => {
+      try {
+        const applicationDomain: ApplicationDomain | undefined = await EpSdkApplicationDomainsService.getById({ applicationDomainId: ApplicationDomainId });
+        expect(applicationDomain, TestLogger.createApiTestFailMessage('applicationDomain === undefined')).to.not.be.undefined;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should delete application domain by id`, async () => {
+      try {
+        const applicationDomain: ApplicationDomain = await EpSdkApplicationDomainsService.deleteById({ applicationDomainId: ApplicationDomainId });
+        expect(applicationDomain, TestLogger.createApiTestFailMessage('applicationDomain === undefined')).to.not.be.undefined;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should create application domain`, async () => {
+      try {
+        const applicationDomainResponse: ApplicationDomainResponse = await ApplicationDomainsService.createApplicationDomain({
+          requestBody: {
+            name: ApplicationDomainName,
+          }
+        });
+        ApplicationDomainId = applicationDomainResponse.data.id;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should delete application domain by name`, async () => {
+      try {
+        const applicationDomain: ApplicationDomain = await EpSdkApplicationDomainsService.deleteByName({ applicationDomainName: ApplicationDomainName });
+        expect(applicationDomain, TestLogger.createApiTestFailMessage('applicationDomain === undefined')).to.not.be.undefined;
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        expect(false, TestLogger.createEpSdkTestFailMessage('failed', e)).to.be.true;
+      }
+    });
+
+    it(`${scriptName}: should catch delete application domain by name that doesn't exist`, async () => {
+      const NonExistentName = 'non-existent-application-domain';
+      try {
+        const applicationDomain: ApplicationDomain = await EpSdkApplicationDomainsService.deleteByName({ applicationDomainName: NonExistentName });
+      } catch(e) {
+        if(e instanceof ApiError) expect(false, TestLogger.createApiTestFailMessage('failed')).to.be.true;
+        expect(e instanceof EpSdkError, TestLogger.createNotEpSdkErrorMesssage(e)).to.be.true;
+        const epSdkServiceError: EpSdkServiceError = e;
+        expect(epSdkServiceError.toString(), TestLogger.createApiTestFailMessage(`error does not contain ${NonExistentName}`)).to.contain(NonExistentName);
+      }
+    });
+
+});
+