+++ ReleaseNotes.md
@@ -2,6 +2,20 @@
 
 Solace Event Portal SDK.
 
+## Version 0.2.0-alpha
+  * [Solace Event Portal OpenAPI](https://github.com/solace-iot-team/ep-sdk/blob/main/resources/sep-openapi-spec.2.0.0-ea.json): '2.0.0-ea'
+
+**New Features:**
+- **Services**
+  - added services framework
+  - added `EpSdkApplicationDomainsService`
+- **Logger**
+  - added `EpSdkLogger` - allows user to initialize the sdk with their own logger
+  - default: `EpSdkConsoleLogger` - uses console to log
+- **Errors**
+  - added `EpSdkErrors` - a collection of sdk specific errors
+
+
 ## Version 0.1.0-alpha
   * [Solace Event Portal OpenAPI](https://github.com/solace-iot-team/ep-sdk/blob/main/resources/sep-openapi-spec.2.0.0-ea.json): '2.0.0-ea'